reorganizing the code
