+++ custom_components/magic_areas/base/primitives.py
@@ -1,0 +1,159 @@
+from datetime import datetime
+
+from homeassistant.components.binary_sensor import BinarySensorEntity
+from homeassistant.components.sensor import SensorEntity
+from homeassistant.const import STATE_ON
+
+from custom_components.magic_areas.base import MagicSensorBase
+from custom_components.magic_areas.const import (
+    CONF_ON_STATES,
+    INVALID_STATES
+)
+
+class BinarySensorBase(MagicSensorBase, BinarySensorEntity):
+    _device_class = None
+    last_off_time = None
+    area = None
+    _state = False
+
+    def __init__(self, area, device_class):
+
+        super().__init__()
+
+        self.area = area
+        self._device_class = device_class
+
+    @property
+    def is_on(self):
+        """Return true if the area is occupied."""
+        return self._state
+
+    def sensor_state_change(self, entity_id, from_state, to_state):
+        self.logger.debug(f"{self.name}: sensor '{entity_id}' changed to {to_state.state}")
+
+        if to_state.state in INVALID_STATES:
+            self.logger.debug(
+                f"{self.name}: sensor '{entity_id}' has invalid state {to_state.state}"
+            )
+            return None
+
+        if to_state and to_state.state not in self.area.config.get(CONF_ON_STATES):
+            self.last_off_time = datetime.utcnow()  # Update last_off_time
+
+        return self._update_state()
+
+    def _get_sensors_state(self, valid_states=[STATE_ON]):
+        self.logger.debug(
+            f"[Area: {self.area.slug}] Updating state. (Valid states: {valid_states})"
+        )
+
+        active_sensors = []
+        active_areas = set()
+
+        # Loop over all entities and check their state
+        for sensor in self.sensors:
+            try:
+                entity = self.hass.states.get(sensor)
+
+                if not entity:
+                    self.logger.info(
+                        f"[Area: {self.area.slug}] Could not get sensor state: {sensor} entity not found, skipping"
+                    )
+                    continue
+
+                self.logger.debug(
+                    f"[Area: {self.area.slug}] Sensor {sensor} state: {entity.state}"
+                )
+
+                # Skip unavailable entities
+                if entity.state in INVALID_STATES:
+                    self.logger.debug(
+                        f"[Area: {self.area.slug}] Sensor '{sensor}' is unavailable, skipping..."
+                    )
+                    continue
+
+                if entity.state in valid_states:
+                    self.logger.debug(
+                        f"[Area: {self.area.slug}] Valid presence sensor found: {sensor}."
+                    )
+                    active_sensors.append(sensor)
+
+            except Exception as e:
+                self.logger.error(
+                    f"[{self.name}] Error getting entity state for '{sensor}': {str(e)}"
+                )
+                pass
+
+        self._attributes["active_sensors"] = active_sensors
+
+        self.logger.debug(f"[Area: {self.area.slug}] Active sensors: {active_sensors}")
+
+        if self.area.is_meta():
+            active_areas = self.area.get_active_areas()
+            self.logger.debug("[Area: {self.area.slug}] Active areas: {active_areas}")
+            self._attributes["active_areas"] = active_areas
+
+        return len(active_sensors) > 0
+    
+
+class SensorBase(MagicSensorBase, SensorEntity):
+    _mode = "mean"
+
+    @property
+    def state(self):
+        """Return the state of the entity"""
+        return self._state
+
+    def sensor_state_change(self, entity_id, from_state, to_state):
+        self.logger.debug(f"{self.name}: sensor '{entity_id}' changed to {to_state.state}")
+
+        if to_state.state in INVALID_STATES:
+            self.logger.debug(
+                f"{self.name}: sensor '{entity_id}' has invalid state {to_state.state}"
+            )
+            return None
+
+        return self._update_state()
+
+    def _get_sensors_state(self):
+        sensor_values = []
+
+        # Loop over all entities and check their state
+        for sensor in self.sensors:
+            try:
+                entity = self.hass.states.get(sensor)
+
+                if not entity:
+                    self.logger.info(
+                        f"[{self.name}] Could not get sensor state: {sensor} entity not found, skipping"
+                    )
+                    continue
+
+                # Skip unavailable entities
+                if entity.state in INVALID_STATES:
+                    continue
+
+            except Exception as e:
+                self.logger.error(
+                    f"[{self.name}] Error getting entity state for '{sensor}': {str(e)}"
+                )
+                continue
+
+            try:
+                sensor_values.append(float(entity.state))
+            except ValueError as e:
+                err_str = str(e)
+                self.logger.info(
+                    f"[{self.name}] Non-numeric sensor value ({err_str}) for entity {entity.entity_id}, skipping"
+                )
+                continue
+
+        ret = 0.0
+
+        if sensor_values:
+            if self._mode == "sum":
+                ret = sum(sensor_values)
+            else:
+                ret = mean(sensor_values)
+
+        return round(ret, 2)
