+++ custom_components/magic_areas/binary_sensor.py
@@ -1,5 +1,3 @@
-DEPENDENCIES = ["magic_areas", "media_player", "binary_sensor"]
-
 import logging
 from datetime import datetime, timedelta
 
@@ -19,8 +17,9 @@ from homeassistant.helpers.event import (
     call_later,
 )
 
-from .base import AggregateBase, BinarySensorBase
-from .const import (
+from custom_components.magic_areas.base import AggregateBase
+from custom_components.magic_areas.base.primitives import BinarySensorBase
+from custom_components.magic_areas.const import (
     AREA_STATE_BRIGHT,
     AREA_STATE_CLEAR,
     AREA_STATE_DARK,
@@ -79,17 +78,16 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
 
 async def load_sensors(hass, async_add_entities, area):
     # Create basic presence sensor
-    async_add_entities([AreaPresenceBinarySensor(hass, area)])
+    async_add_entities([AreaPresenceBinarySensor(area)])
 
     # Create extra sensors
     if area.has_feature(CONF_FEATURE_AGGREGATION):
-        await create_aggregate_sensors(hass, area, async_add_entities)
+        await create_aggregate_sensors(area, async_add_entities)
 
     if area.has_feature(CONF_FEATURE_HEALTH):
-        await create_health_sensors(hass, area, async_add_entities)
+        await create_health_sensors(area, async_add_entities)
 
-
-async def create_health_sensors(hass, area, async_add_entities):
+async def create_health_sensors(area, async_add_entities):
     if not area.has_feature(CONF_FEATURE_HEALTH):
         return
 
@@ -113,10 +111,10 @@ async def create_health_sensors(hass, area, async_add_entities):
         return
 
     _LOGGER.debug(f"Creating health sensor for area ({area.slug})")
-    async_add_entities([AreaDistressBinarySensor(hass, area)])
+    async_add_entities([AreaDistressBinarySensor(area)])
 
 
-async def create_aggregate_sensors(hass, area, async_add_entities):
+async def create_aggregate_sensors(area, async_add_entities):
     # Create aggregates
     if not area.has_feature(CONF_FEATURE_AGGREGATION):
         return
@@ -147,24 +145,21 @@ async def create_aggregate_sensors(hass, area, async_add_entities):
         _LOGGER.debug(
             f"Creating aggregate sensor for device_class '{device_class}' with {entity_count} entities ({area.slug})"
         )
-        aggregates.append(AreaSensorGroupBinarySensor(hass, area, device_class))
+        aggregates.append(AreaSensorGroupBinarySensor(area, device_class))
 
     async_add_entities(aggregates)
 
 
 class AreaPresenceBinarySensor(BinarySensorBase):
-    def __init__(self, hass, area):
+    def __init__(self, area):
         """Initialize the area presence binary sensor."""
 
-        self.area = area
-        self.hass = hass
+        super().__init__(area, device_class = BinarySensorDeviceClass.OCCUPANCY)
+
         self._name = f"Area ({self.area.name})"
-        self.area.occupied = False
+
         self.last_off_time = datetime.utcnow()
         self.clear_timeout_callback = None
-
-        self._device_class = BinarySensorDeviceClass.OCCUPANCY
-        self.sensors = []
 
     def load_presence_sensors(self) -> None:
         if self.area.is_meta():
@@ -262,7 +257,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
         is_new_entry = last_state is None  # newly added to HA
 
         if is_new_entry:
-            _LOGGER.debug(f"New sensor created: {self.name}")
+            self.logger.debug(f"New sensor created: {self.name}")
             self._update_state()
         else:
             _LOGGER.debug(f"Sensor {self.name} restored [state={last_state.state}]")
@@ -273,7 +268,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
             self.schedule_update_ha_state()
 
     async def _initialize(self, _=None) -> None:
-        _LOGGER.debug(f"{self.name} Sensor initializing.")
+        self.logger.debug(f"{self.name} Sensor initializing.")
 
         self.load_presence_sensors()
         self.load_attributes()
@@ -284,9 +279,9 @@ class AreaPresenceBinarySensor(BinarySensorBase):
         _LOGGER.debug(f"{self.name} Sensor initialized.")
 
     async def _setup_listeners(self, _=None) -> None:
-        _LOGGER.debug("%s: Called '_setup_listeners'", self.name)
+        self.logger.debug("%s: Called '_setup_listeners'", self.name)
         if not self.hass.is_running:
-            _LOGGER.debug("%s: Cancelled '_setup_listeners'", self.name)
+            self.logger.debug("%s: Cancelled '_setup_listeners'", self.name)
             return
 
         # Track presence sensors
@@ -308,7 +303,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
             if not tracked_entity:
                 continue
 
-            _LOGGER.debug(f"Secondary state tracking: {tracked_entity}")
+            self.logger.debug(f"Secondary state tracking: {tracked_entity}")
 
             self.async_on_remove(
                 async_track_state_change(
@@ -323,12 +318,12 @@ class AreaPresenceBinarySensor(BinarySensorBase):
         )
 
     def secondary_state_change(self, entity_id, from_state, to_state):
-        _LOGGER.debug(
+        self.logger.debug(
             f"{self.name}: Secondary state change: entity '{entity_id}' changed to {to_state.state}"
         )
 
         if to_state.state in INVALID_STATES:
-            _LOGGER.debug(
+            self.logger.debug(
                 f"{self.name}: sensor '{entity_id}' has invalid state {to_state.state}"
             )
             return None
@@ -346,7 +341,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
         # Calculate what's new
         new_states = current_state - last_state
         lost_states = last_state - current_state
-        _LOGGER.debug(
+        self.logger.debug(
             f"{self.name}: Current state: {current_state}, last state: {last_state} -> new states {new_states} / lost states {lost_states}"
         )
 
@@ -387,7 +382,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
         states.append(AREA_STATE_OCCUPIED if current_state else AREA_STATE_CLEAR)
         if current_state != last_state:
             self.area.last_changed = datetime.utcnow()
-            _LOGGER.debug(
+            self.logger.debug(
                 f"{self.area.name}: State changed to {current_state} at {self.area.last_changed}"
             )
 
@@ -427,7 +422,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
             entity = self.hass.states.get(secondary_state_entity)
 
             if entity.state.lower() == secondary_state_value.lower():
-                _LOGGER.debug(
+                self.logger.debug(
                     f"{self.area.name}: Secondary state: {secondary_state_entity} is at {secondary_state_value}, adding {configurable_state}"
                 )
                 states.append(configurable_state)
@@ -464,7 +459,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
 
         timeout = self.get_clear_timeout()
 
-        _LOGGER.debug(f"{self.area.name}: Scheduling clear in {timeout} seconds")
+        self.logger.debug(f"{self.area.name}: Scheduling clear in {timeout} seconds")
         self.clear_timeout_callback = call_later(
             self.hass, timeout, self.refresh_states
         )
@@ -490,7 +485,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
         time_now = datetime.utcnow()
 
         if time_now >= clear_time:
-            _LOGGER.debug(f"{self.area.name}: Clear Timeout exceeded.")
+            self.logger.debug(f"{self.area.name}: Clear Timeout exceeded.")
             self.remove_clear_timeout()
             return True
 
@@ -507,12 +502,12 @@ class AreaPresenceBinarySensor(BinarySensorBase):
                 return False
 
             if self.is_on_clear_timeout():
-                _LOGGER.debug(f"{self.area.name}: Area is on timeout")
+                self.logger.debug(f"{self.area.name}: Area is on timeout")
                 if self.timeout_exceeded():
                     return False
             else:
                 if self.area.is_occupied() and not area_state:
-                    _LOGGER.debug(
+                    self.logger.debug(
                         f"{self.area.name}: Area not on timeout, setting call_later"
                     )
                     self.set_clear_timeout()
@@ -529,7 +524,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
             state in new_states for state in [AREA_STATE_OCCUPIED, AREA_STATE_CLEAR]
         )
 
-        _LOGGER.debug(
+        self.logger.debug(
             f"{self.area.name}: States updated. New states: {new_states} / Lost states: {lost_states}"
         )
 
@@ -544,7 +539,7 @@ class AreaPresenceBinarySensor(BinarySensorBase):
 
     def report_state_change(self, states_tuple=([], [])):
         new_states, lost_states = states_tuple
-        _LOGGER.debug(
+        self.logger.debug(
             f"Reporting state change for {self.area.name} (new states: {new_states}/lost states: {lost_states})"
         )
         dispatcher_send(
@@ -553,48 +548,42 @@ class AreaPresenceBinarySensor(BinarySensorBase):
 
 
 class AreaSensorGroupBinarySensor(BinarySensorBase, AggregateBase):
-    def __init__(self, hass, area, device_class):
+    def __init__(self, area, device_class):
         """Initialize an area sensor group binary sensor."""
 
-        self.area = area
-        self.hass = hass
-        self._device_class = device_class
-        self._state = False
+        super().__init__(area, device_class)
 
         device_class_name = " ".join(device_class.split("_")).title()
         self._name = f"Area {device_class_name} ({self.area.name})"
 
     async def _initialize(self, _=None) -> None:
-        _LOGGER.debug(f"{self.name} Sensor initializing.")
+        self.logger.debug(f"{self.name} Sensor initializing.")
 
         self.load_sensors(BINARY_SENSOR_DOMAIN)
 
         # Setup the listeners
         await self._setup_listeners()
 
-        _LOGGER.debug(f"{self.name} Sensor initialized.")
+        self.logger.debug(f"{self.name} Sensor initialized.")
 
 
 class AreaDistressBinarySensor(BinarySensorBase, AggregateBase):
-    def __init__(self, hass, area):
+    def __init__(self, area):
         """Initialize an area sensor group binary sensor."""
 
-        self.area = area
-        self.hass = hass
-        self._device_class = BinarySensorDeviceClass.PROBLEM
-        self._state = False
+        super().__init__(area, device_class=BinarySensorDeviceClass.PROBLEM)
 
         self._name = f"Area Health ({self.area.name})"
 
     async def _initialize(self, _=None) -> None:
-        _LOGGER.debug(f"{self.name} Sensor initializing.")
+        self.logger.debug(f"{self.name} Sensor initializing.")
 
         self.load_sensors()
 
         # Setup the listeners
         await self._setup_listeners()
 
-        _LOGGER.debug(f"{self.name} Sensor initialized.")
+        self.logger.debug(f"{self.name} Sensor initialized.")
 
     def load_sensors(self):
         # Fetch sensors
